portal/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.23
pydantic==2.5.3
pydantic-settings==2.0.3
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.26.0
apscheduler==3.10.4

